# docs/stridehub/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StrideHub Training API",
    "version": "1.0.0",
    "description": "Single-athlete, self-hosted training API. PUT merges supplied fields. Existing record mutations require expected version. Dates in queries use runner profile timezone. Never retry a 409 by blindly replacing the version."
  },
  "servers": [
    {
      "url": "/api/v1/training"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/sessions/batch": {
      "put": {
        "summary": "Atomically create/update up to 100 sessions; version 0 creates, any failure rolls back all",
        "operationId": "batchUpsertSessions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sessions"
                ],
                "properties": {
                  "sessions": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 100,
                    "items": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/SessionUpdate"
                        },
                        {
                          "required": [
                            "id",
                            "version"
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Session"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict, duplicate or referenced resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "428": {
            "description": "Expected version required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Dreeve drv_ token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Session": {
        "type": "object",
        "required": [
          "title",
          "startAt",
          "durationMinutes",
          "type"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "startAt": {
            "type": "string",
            "format": "date-time",
            "description": "RFC3339 timestamp with explicit UTC offset"
          },
          "durationMinutes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1440
          },
          "distanceKm": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "maximum": 1000
          },
          "type": {
            "type": "string",
            "enum": [
              "easy",
              "long",
              "tempo",
              "interval",
              "recovery",
              "race",
              "rest"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "planned",
              "completed",
              "skipped",
              "cancelled"
            ],
            "default": "planned"
          },
          "raceId": {
            "type": [
              "string",
              "null"
            ]
          },
          "location": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "required": [
                  "label",
                  "latitude",
                  "longitude"
                ],
                "additionalProperties": false,
                "properties": {
                  "label": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "latitude": {
                    "type": "number",
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": -180,
                    "maximum": 180
                  }
                }
              }
            ]
          },
          "steps": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "object",
              "required": [
                "kind",
                "minutes"
              ],
              "properties": {
                "kind": {
                  "type": "string",
                  "maxLength": 80
                },
                "minutes": {
                  "type": "number",
                  "minimum": 0.1,
                  "maximum": 1440
                },
                "distanceKm": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 1000
                },
                "target": {
                  "type": "string",
                  "maxLength": 200
                }
              }
            }
          },
          "fuelPlan": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "object",
              "required": [
                "minute",
                "item"
              ],
              "properties": {
                "minute": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 1440
                },
                "item": {
                  "type": "string",
                  "maxLength": 200
                },
                "carbsGrams": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 1000
                },
                "fluidMl": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 10000
                }
              }
            }
          },
          "notes": {
            "type": "string",
            "maxLength": 4000
          },
          "activityId": {
            "type": [
              "string",
              "null"
            ]
          },
          "feedback": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "properties": {
                  "rpe": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "thermalFeeling": {
                    "type": "string",
                    "enum": [
                      "cold",
                      "comfortable",
                      "hot"
                    ]
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 4000
                  }
                }
              }
            ]
          },
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,80}$"
          },
          "version": {
            "type": "integer",
            "minimum": 0,
            "description": "Expected current version on mutations. Zero only for initial creation."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
      "SessionUpdate": {
        "type": "object",
        "required": [
          "version"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "startAt": {
            "type": "string",
            "format": "date-time",
            "description": "RFC3339 timestamp with explicit UTC offset"
          },
          "durationMinutes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1440
          },
          "distanceKm": {
            "type": [
              "number",
              "null"
            ],
            "minimum": 0,
            "maximum": 1000
          },
          "type": {
            "type": "string",
            "enum": [
              "easy",
              "long",
              "tempo",
              "interval",
              "recovery",
              "race",
              "rest"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "planned",
              "completed",
              "skipped",
              "cancelled"
            ],
            "default": "planned"
          },
          "raceId": {
            "type": [
              "string",
              "null"
            ]
          },
          "location": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "required": [
                  "label",
                  "latitude",
                  "longitude"
                ],
                "additionalProperties": false,
                "properties": {
                  "label": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "latitude": {
                    "type": "number",
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": -180,
                    "maximum": 180
                  }
                }
              }
            ]
          },
          "steps": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "object",
              "required": [
                "kind",
                "minutes"
              ],
              "properties": {
                "kind": {
                  "type": "string",
                  "maxLength": 80
                },
                "minutes": {
                  "type": "number",
                  "minimum": 0.1,
                  "maximum": 1440
                },
                "distanceKm": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 1000
                },
                "target": {
                  "type": "string",
                  "maxLength": 200
                }
              }
            }
          },
          "fuelPlan": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "object",
              "required": [
                "minute",
                "item"
              ],
              "properties": {
                "minute": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 1440
                },
                "item": {
                  "type": "string",
                  "maxLength": 200
                },
                "carbsGrams": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 1000
                },
                "fluidMl": {
                  "type": "number",
                  "minimum": 0,
                  "maximum": 10000
                }
              }
            }
          },
          "notes": {
            "type": "string",
            "maxLength": 4000
          },
          "activityId": {
            "type": [
              "string",
              "null"
            ]
          },
          "feedback": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "properties": {
                  "rpe": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "thermalFeeling": {
                    "type": "string",
                    "enum": [
                      "cold",
                      "comfortable",
                      "hot"
                    ]
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 4000
                  }
                }
              }
            ]
          },
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,80}$"
          },
          "version": {
            "type": "integer",
            "minimum": 0,
            "description": "Expected current version on mutations. Zero only for initial creation."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
    }
  }
}